requirements.txt
# Core dependencies
PyQt6>=6.5.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
SpeechRecognition>=3.10.0
sounddevice>=0.4.6
numpy>=1.24.0
pyttsx3>=2.90
webrtcvad>=2.0.10

# Advanced capabilities
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Code validation
pytesseract>=0.3.10
mss>=9.0.1

# Audio processing
pyaudio>=0.2.14
openai-whisper>=20231117  # Optional: for advanced transcription

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
